+++ yellowbox/extras/azure_storage.py
@@ -3,11 +3,13 @@ Azure Blob Storage module, for creating container, uploading files to it and dow
 """
 from __future__ import annotations
 
-from typing import Union
+from typing import List, Optional, Sequence, Union
 
 from docker import DockerClient
+from docker.api import container
+from docker.models.networks import Network
 
-from yellowbox.containers import create_and_pull, get_ports
+from yellowbox.containers import create_and_pull, get_ports, short_id
 from yellowbox.subclasses import RunMixin, SingleContainerService
 from yellowbox.utils import retry
 
@@ -48,11 +50,22 @@ class BlobStorageService(SingleContainerService, RunMixin):
 
     @property
     def connection_string(self):
+        """Connection string to connect from host to container"""
         return (
             f"DefaultEndpointsProtocol=http;"
             f"AccountName={self.account_name};"
             f"AccountKey={self.account_key};"
             f"BlobEndpoint=http://localhost:{self.client_port()}/{self.account_name};")
+
+    @property
+    def container_connection_string(self):
+        """Connection string to connect across containers in the same network"""
+        return (
+            f"DefaultEndpointsProtocol=http;"
+            f"AccountName={self.account_name};"
+            f"AccountKey={self.account_key};"
+            f"BlobEndpoint="
+            f"http://{short_id(self.container)}:{BLOB_STORAGE_DEFAULT_PORT}/{self.account_name};")
 
     def start(self):
         super().start()
@@ -63,3 +76,11 @@ class BlobStorageService(SingleContainerService, RunMixin):
 
         retry(check_ready, _ResourceNotReady)
         return self
+
+    def connect(self, network: Network, aliases: Optional[List[str]] = None,
+                **kwargs) -> Sequence[str]:
+        # Make sure the id is in the aliases list. Needed for the container
+        # connection string.
+        if aliases is not None:
+            aliases.append(short_id(self.container))
+        return super().connect(network, aliases=aliases, **kwargs)